requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
asyncio-throttle==1.0.2
python-dotenv==1.0.0